src/data_embedding/requirements.txt
fastapi==0.120.0
langchain_openai==1.0.1
psycopg2_binary==2.9.10
pyodbc==5.2.0
python-dotenv==1.1.1
Requests==2.32.5
sentence_transformers==5.1.1
uvicorn==0.38.0
gunicorn